refactor: Use recoil selector instead of function.
